+++ src/realm/group.cpp
@@ -326,13 +326,12 @@ void Transaction::upgrade_file_format(int target_file_format_version)
             m_top.add(initial_history_schema_version); // Throws
         }
         set_file_format_version(7);
-        commit_and_continue_as_read();
-        promote_to_write();
+        commit_and_continue_writing();
     }
 
     // NOTE: Additional future upgrade steps go here.
     if (current_file_format_version <= 9 && target_file_format_version >= 10) {
-        // DisableReplication disable_replication(*this);
+        DisableReplication disable_replication(*this);
 
         std::vector<TableKey> table_keys;
         for (size_t t = 0; t < m_table_names.size(); t++) {
@@ -341,10 +340,7 @@ void Transaction::upgrade_file_format(int target_file_format_version)
             table_keys.push_back(table->get_key());
         }
 
-        auto commit_and_continue = [this]() {
-            commit_and_continue_as_read();
-            promote_to_write();
-        };
+        auto commit_and_continue = [this]() { commit_and_continue_writing(); };
         for (auto k : table_keys) {
             get_table(k)->migrate_column_info(commit_and_continue);
         }
@@ -915,9 +911,9 @@ void Group::rename_table(TableKey key, StringData new_name, bool require_unique_
 
 class Group::DefaultTableWriter : public Group::TableWriter {
 public:
-    DefaultTableWriter(const Group& group, bool write_history)
+    DefaultTableWriter(const Group& group, bool should_write_history)
         : m_group(group)
-        , m_write_history(write_history)
+        , m_should_write_history(should_write_history)
     {
     }
     ref_type write_names(_impl::OutputStream& out) override
@@ -946,7 +942,7 @@ public:
                                                              m_group.m_top.get_ref(), version, history_type,
                                                              history_schema_version);
             REALM_ASSERT(history_type != Replication::hist_None);
-            if (!m_write_history ||
+            if (!m_should_write_history ||
                 (history_type != Replication::hist_SyncClient && history_type != Replication::hist_SyncServer)) {
                 return info; // Only sync history should be preserved when writing to a new file
             }
@@ -963,7 +959,7 @@ public:
 
 private:
     const Group& m_group;
-    bool m_write_history;
+    bool m_should_write_history;
 };
 
 void Group::write(std::ostream& out, bool pad) const